% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin_addPeer}
\alias{admin_addPeer}
\title{New node based on the enode}
\usage{
admin_addPeer(enode)
}
\arguments{
\item{enode}{String - URL of the remote peer to start tracking.}
}
\value{
Boolean - \code{true} the peer was accepted for tracking or
  \code{false} if some error occurred.
}
\description{
\code{admin_addPeer} request to add a new remote node to the list of tracked
static nodes.
}
\examples{
\donttest{
admin_addPeer('enode://c1a07558238c0b31657450dd34a558752d63150ce334f3e99b9187
262b612f48a713a083cd1601bfe3bba761a908264320885633fa81d6d6ca0ef7a6e84a2bcd
  @[127.0.0.1]:30301') }

}
\seealso{
Other admin functions: \code{\link{admin_datadir}},
  \code{\link{admin_nodeInfo}}, \code{\link{admin_peers}},
  \code{\link{admin_setSolc}},
  \code{\link{admin_startRPC}},
  \code{\link{admin_startWS}}, \code{\link{admin_stopRPC}},
  \code{\link{admin_stopWS}}, \code{\link{gethr}}
}
\concept{admin functions}
