% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin_datadir}
\alias{admin_datadir}
\title{Path being used by the node.}
\usage{
admin_datadir()
}
\value{
String - Absolute path being used by the node.
}
\description{
\code{admin_datadir} returns the absolute path the running Geth node.
}
\examples{
\donttest{
admin_datadir()
}

}
\seealso{
Other admin functions: \code{\link{admin_addPeer}},
  \code{\link{admin_nodeInfo}}, \code{\link{admin_peers}},
  \code{\link{admin_setSolc}},
  \code{\link{admin_startRPC}},
  \code{\link{admin_startWS}}, \code{\link{admin_stopRPC}},
  \code{\link{admin_stopWS}}
}
\concept{admin functions}
