% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin_nodeInfo}
\alias{admin_nodeInfo}
\title{Node basic information}
\usage{
admin_nodeInfo()
}
\value{
Object - Properties with information about the node (enode, version, port, ip, id, etc.).
}
\description{
\code{admin_nodeInfo} returns all the information known about the running Geth node at the networking granularity.
}
\details{
These include general information about the node itself as a participant of the P2P overlay protocol, as well as specialized information added by each of the running application protocols (e.g. eth, les, shh, bzz).
}
\examples{
\donttest{
admin_nodeInfo()
}

}
\seealso{
Other admin functions: \code{\link{admin_addPeer}},
  \code{\link{admin_datadir}}, \code{\link{admin_peers}},
  \code{\link{admin_setSolc}},
  \code{\link{admin_startRPC}},
  \code{\link{admin_startWS}}, \code{\link{admin_stopRPC}},
  \code{\link{admin_stopWS}}
}
\concept{admin functions}
