% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin_peers}
\alias{admin_peers}
\title{Peers basic information}
\usage{
admin_peers()
}
\value{
Object - Properties with information about the peers.
}
\description{
\code{admin_peers} returns all the information known about the connected
remote nodes at the networking granularity.
}
\details{
These include general information about the nodes themselves as participants
of the P2P overlay protocol, as well as specialized information added by each
of the running application protocols (e.g. eth, les, shh, bzz).
}
\examples{
\donttest{
admin_peers()
}

}
\seealso{
Other admin functions: \code{\link{admin_addPeer}},
  \code{\link{admin_datadir}},
  \code{\link{admin_nodeInfo}},
  \code{\link{admin_setSolc}},
  \code{\link{admin_startRPC}},
  \code{\link{admin_startWS}}, \code{\link{admin_stopRPC}},
  \code{\link{admin_stopWS}}, \code{\link{gethr}}
}
\concept{admin functions}
