% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin_setSolc}
\alias{admin_setSolc}
\title{Solidity compiler path}
\usage{
admin_setSolc(path)
}
\arguments{
\item{path}{String - Solidity compiler path defaults to /usr/bin/solc if not set, so you only need to change it for using a non-standard compiler location.}
}
\description{
\code{admin_setSolc} sets the Solidity compiler path to be used by the node when invoking the eth_compileSolidity RPC method.
}
\examples{
\donttest{
admin_setSolc("/usr/bin/solidityc")
}

}
\seealso{
Other admin functions: \code{\link{admin_addPeer}},
  \code{\link{admin_datadir}},
  \code{\link{admin_nodeInfo}}, \code{\link{admin_peers}},
  \code{\link{admin_startRPC}},
  \code{\link{admin_startWS}}, \code{\link{admin_stopRPC}},
  \code{\link{admin_stopWS}}
}
\concept{admin functions}
