% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin_stopRPC}
\alias{admin_stopRPC}
\title{HTTP based JSON RPC API stop.}
\usage{
admin_stopRPC()
}
\value{
Boolean - \code{true} the execution was successful.
}
\description{
\code{admin_stopRPC} stops the HTTP based JSON RPC API webserver.
}
\examples{
\donttest{
admin_stopRPC()
}

}
\seealso{
Other admin functions: \code{\link{admin_addPeer}},
  \code{\link{admin_datadir}},
  \code{\link{admin_nodeInfo}}, \code{\link{admin_peers}},
  \code{\link{admin_setSolc}},
  \code{\link{admin_startRPC}},
  \code{\link{admin_startWS}}, \code{\link{admin_stopWS}}
}
\concept{admin functions}
