% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{debug_backtraceAt}
\alias{debug_backtraceAt}
\title{Logging backtrace location.}
\usage{
debug_backtraceAt(path)
}
\arguments{
\item{path}{String - Backtrace location.}
}
\description{
\code{debug_backtraceAt} sets the logging backtrace location. When a
backtrace location is set and a log message is emitted at that location, the
stack of the goroutine executing the log statement will be printed to stderr.
}
\examples{
\donttest{
debug_backtraceAt('my_server.go:443')
}

}
\seealso{
Other debug functions: \code{\link{debug_blockProfile}},
  \code{\link{debug_cpuProfile}},
  \code{\link{debug_dumpBlock}},
  \code{\link{debug_gcStats}},
  \code{\link{debug_getBlockRlp}},
  \code{\link{debug_goTrace}},
  \code{\link{debug_memStats}},
  \code{\link{debug_seedHash}},
  \code{\link{debug_setBlockProfileRate}},
  \code{\link{debug_setHead}}, \code{\link{debug_stacks}},
  \code{\link{debug_startCPUProfile}},
  \code{\link{debug_startGoTrace}},
  \code{\link{debug_stopCPUProfile}},
  \code{\link{debug_stopGoTrace}},
  \code{\link{debug_traceBlockByHash}},
  \code{\link{debug_traceBlockByNumber}},
  \code{\link{debug_traceBlockFromFile}},
  \code{\link{debug_traceBlock}},
  \code{\link{debug_traceTransaction}},
  \code{\link{debug_verbosity}},
  \code{\link{debug_vmodule}},
  \code{\link{debug_writeBlockProfile}},
  \code{\link{debug_writeMemProfile}}
}
\concept{debug functions}
