% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{debug_cpuProfile}
\alias{debug_cpuProfile}
\title{CPU profiling activation.}
\usage{
debug_cpuProfile(file, seconds)
}
\arguments{
\item{file}{String - File to write the data.}

\item{seconds}{Integer - Seconds to write the data.}
}
\description{
\code{debug_cpuProfile} turns on CPU profiling for the given duration and
writes profile data to disk.
}
\examples{
\donttest{
debug_cpuProfile('file.log', 5)
}

}
\seealso{
Other debug functions: \code{\link{debug_backtraceAt}},
  \code{\link{debug_blockProfile}},
  \code{\link{debug_dumpBlock}},
  \code{\link{debug_gcStats}},
  \code{\link{debug_getBlockRlp}},
  \code{\link{debug_goTrace}},
  \code{\link{debug_memStats}},
  \code{\link{debug_seedHash}},
  \code{\link{debug_setBlockProfileRate}},
  \code{\link{debug_setHead}}, \code{\link{debug_stacks}},
  \code{\link{debug_startCPUProfile}},
  \code{\link{debug_startGoTrace}},
  \code{\link{debug_stopCPUProfile}},
  \code{\link{debug_stopGoTrace}},
  \code{\link{debug_traceBlockByHash}},
  \code{\link{debug_traceBlockByNumber}},
  \code{\link{debug_traceBlockFromFile}},
  \code{\link{debug_traceBlock}},
  \code{\link{debug_traceTransaction}},
  \code{\link{debug_verbosity}},
  \code{\link{debug_vmodule}},
  \code{\link{debug_writeBlockProfile}},
  \code{\link{debug_writeMemProfile}}
}
\concept{debug functions}
