% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{debug_dumpBlock}
\alias{debug_dumpBlock}
\title{State of the block.}
\usage{
debug_dumpBlock(number)
}
\arguments{
\item{number}{Integer - Number of the block.}
}
\value{
Object - Information of the state (balance, code, hash, nonce, root and storage).
}
\description{
\code{debug_dumpBlock} retrieves the state that corresponds to the block number and returns a list of accounts (including storage and code).
}
\examples{
\donttest{
debug_dumpBlock(15)
}

}
\seealso{
Other debug functions: \code{\link{debug_backtraceAt}},
  \code{\link{debug_blockProfile}},
  \code{\link{debug_cpuProfile}},
  \code{\link{debug_gcStats}},
  \code{\link{debug_getBlockRlp}},
  \code{\link{debug_goTrace}},
  \code{\link{debug_memStats}},
  \code{\link{debug_seedHash}},
  \code{\link{debug_setBlockProfileRate}},
  \code{\link{debug_setHead}}, \code{\link{debug_stacks}},
  \code{\link{debug_startCPUProfile}},
  \code{\link{debug_startGoTrace}},
  \code{\link{debug_stopCPUProfile}},
  \code{\link{debug_stopGoTrace}},
  \code{\link{debug_traceBlockByHash}},
  \code{\link{debug_traceBlockByNumber}},
  \code{\link{debug_traceBlockFromFile}},
  \code{\link{debug_traceBlock}},
  \code{\link{debug_traceTransaction}},
  \code{\link{debug_verbosity}},
  \code{\link{debug_vmodule}},
  \code{\link{debug_writeBlockProfile}},
  \code{\link{debug_writeMemProfile}}
}
\concept{debug functions}
