% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{debug_getBlockRlp}
\alias{debug_getBlockRlp}
\title{Recursive Length Prefix of the block.}
\usage{
debug_getBlockRlp(number)
}
\arguments{
\item{number}{Integer - Number of the block.}
}
\value{
Data - Recursive Length Prefix of the block.
}
\description{
\code{debug_getBlockRlp} retrieves and returns the Recursive Length Prefix by
number of block.
}
\examples{
\donttest{
debug_getBlockRlp(29)
}

}
\seealso{
Other debug functions: \code{\link{debug_backtraceAt}},
  \code{\link{debug_blockProfile}},
  \code{\link{debug_cpuProfile}},
  \code{\link{debug_dumpBlock}},
  \code{\link{debug_gcStats}}, \code{\link{debug_goTrace}},
  \code{\link{debug_memStats}},
  \code{\link{debug_seedHash}},
  \code{\link{debug_setBlockProfileRate}},
  \code{\link{debug_setHead}}, \code{\link{debug_stacks}},
  \code{\link{debug_startCPUProfile}},
  \code{\link{debug_startGoTrace}},
  \code{\link{debug_stopCPUProfile}},
  \code{\link{debug_stopGoTrace}},
  \code{\link{debug_traceBlockByHash}},
  \code{\link{debug_traceBlockByNumber}},
  \code{\link{debug_traceBlockFromFile}},
  \code{\link{debug_traceBlock}},
  \code{\link{debug_traceTransaction}},
  \code{\link{debug_verbosity}},
  \code{\link{debug_vmodule}},
  \code{\link{debug_writeBlockProfile}},
  \code{\link{debug_writeMemProfile}}
}
\concept{debug functions}
