% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{debug_seedHash}
\alias{debug_seedHash}
\title{Seed hash of the block}
\usage{
debug_seedHash(number)
}
\arguments{
\item{number}{Integer - Number of the block.}
}
\value{
Data - Seed hash of the block by number.
}
\description{
\code{debug_seedHash} fetches and retrieves the seed hash of the block by number.
}
\examples{
\donttest{
debug_seedHash(29)
}

}
\seealso{
Other debug functions: \code{\link{debug_backtraceAt}},
  \code{\link{debug_blockProfile}},
  \code{\link{debug_cpuProfile}},
  \code{\link{debug_dumpBlock}},
  \code{\link{debug_gcStats}},
  \code{\link{debug_getBlockRlp}},
  \code{\link{debug_goTrace}},
  \code{\link{debug_memStats}},
  \code{\link{debug_setBlockProfileRate}},
  \code{\link{debug_setHead}}, \code{\link{debug_stacks}},
  \code{\link{debug_startCPUProfile}},
  \code{\link{debug_startGoTrace}},
  \code{\link{debug_stopCPUProfile}},
  \code{\link{debug_stopGoTrace}},
  \code{\link{debug_traceBlockByHash}},
  \code{\link{debug_traceBlockByNumber}},
  \code{\link{debug_traceBlockFromFile}},
  \code{\link{debug_traceBlock}},
  \code{\link{debug_traceTransaction}},
  \code{\link{debug_verbosity}},
  \code{\link{debug_vmodule}},
  \code{\link{debug_writeBlockProfile}},
  \code{\link{debug_writeMemProfile}}
}
\concept{debug functions}
