% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{debug_setBlockProfileRate}
\alias{debug_setBlockProfileRate}
\title{Block profile data collection rate.}
\usage{
debug_setBlockProfileRate(rate)
}
\arguments{
\item{rate}{Integer - Rate of the profiling. A non-zero rate enables block
profiling, setting it to zero stops the profile.}
}
\description{
\code{debug_setBlockProfileRate} sets the rate (in samples/sec) of goroutine
block profile data collection.
}
\examples{
\donttest{
debug_setBlockProfileRate(100)
}

}
\seealso{
Other debug functions: \code{\link{debug_backtraceAt}},
  \code{\link{debug_blockProfile}},
  \code{\link{debug_cpuProfile}},
  \code{\link{debug_dumpBlock}},
  \code{\link{debug_gcStats}},
  \code{\link{debug_getBlockRlp}},
  \code{\link{debug_goTrace}},
  \code{\link{debug_memStats}},
  \code{\link{debug_seedHash}},
  \code{\link{debug_setHead}}, \code{\link{debug_stacks}},
  \code{\link{debug_startCPUProfile}},
  \code{\link{debug_startGoTrace}},
  \code{\link{debug_stopCPUProfile}},
  \code{\link{debug_stopGoTrace}},
  \code{\link{debug_traceBlockByHash}},
  \code{\link{debug_traceBlockByNumber}},
  \code{\link{debug_traceBlockFromFile}},
  \code{\link{debug_traceBlock}},
  \code{\link{debug_traceTransaction}},
  \code{\link{debug_verbosity}},
  \code{\link{debug_vmodule}},
  \code{\link{debug_writeBlockProfile}},
  \code{\link{debug_writeMemProfile}}, \code{\link{gethr}}
}
\concept{debug functions}
