% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{debug_vmodule}
\alias{debug_vmodule}
\title{Logging verbosity by pattern.}
\usage{
debug_vmodule(pattern)
}
\arguments{
\item{pattern}{String - Pattern used to indicate the verbosity.}
}
\description{
\code{debug_vmodule} sets the logging verbosity pattern.
}
\examples{
\donttest{
debug_vmodule('eth/*=4')
debug_vmodule('p2p=5')
}

}
\seealso{
Other debug functions: \code{\link{debug_backtraceAt}},
  \code{\link{debug_blockProfile}},
  \code{\link{debug_cpuProfile}},
  \code{\link{debug_dumpBlock}},
  \code{\link{debug_gcStats}},
  \code{\link{debug_getBlockRlp}},
  \code{\link{debug_goTrace}},
  \code{\link{debug_memStats}},
  \code{\link{debug_seedHash}},
  \code{\link{debug_setBlockProfileRate}},
  \code{\link{debug_setHead}}, \code{\link{debug_stacks}},
  \code{\link{debug_startCPUProfile}},
  \code{\link{debug_startGoTrace}},
  \code{\link{debug_stopCPUProfile}},
  \code{\link{debug_stopGoTrace}},
  \code{\link{debug_traceBlockByHash}},
  \code{\link{debug_traceBlockByNumber}},
  \code{\link{debug_traceBlockFromFile}},
  \code{\link{debug_traceBlock}},
  \code{\link{debug_traceTransaction}},
  \code{\link{debug_verbosity}},
  \code{\link{debug_writeBlockProfile}},
  \code{\link{debug_writeMemProfile}}, \code{\link{gethr}}
}
\concept{debug functions}
