% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dec_to_hex}
\alias{dec_to_hex}
\title{Decimal to hexadecimal conversion.}
\usage{
dec_to_hex(number)
}
\arguments{
\item{number}{Integer - Value in decimal.}
}
\value{
Data - Value in hexadecimal.
}
\description{
\code{dec_to_hex} returns the value in hexadecimal.
}
\examples{
dec_to_hex(50)

}
\seealso{
Other utils functions: \code{\link{get_post}},
  \code{\link{get_rpc_address}}, \code{\link{hex_to_dec}},
  \code{\link{hex_to_text}}, \code{\link{is.wholenumber}},
  \code{\link{process_block}}, \code{\link{process_log}},
  \code{\link{process_receipt}},
  \code{\link{process_transaction}},
  \code{\link{set_rpc_address}}, \code{\link{text_to_hex}}
}
\concept{utils functions}
