% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ether.R
\name{ether.toEther}
\alias{ether.toEther}
\title{Conversion to Ether.}
\usage{
ether.toEther(amount, type)
}
\arguments{
\item{amount}{Integer - Amount of the cryptocurrency to convert from.}

\item{type}{String - Unit of the cryptocurrency to convert from: wei, kwei,
mwei, gwei, szabo, finney, ether, kether, mether, gether or tether.}
}
\value{
Integer - Value in Ether.
}
\description{
\code{ether.toEther} returns the value of the cryptocurrency in Ether.
}
\examples{
ether.toEther(50, 'wei')

}
\seealso{
Other ether functions: \code{\link{ether.toFinney}},
  \code{\link{ether.toGether}}, \code{\link{ether.toGwei}},
  \code{\link{ether.toKether}}, \code{\link{ether.toKwei}},
  \code{\link{ether.toMether}}, \code{\link{ether.toMwei}},
  \code{\link{ether.toSzabo}},
  \code{\link{ether.toTether}}, \code{\link{ether.toWei}},
  \code{\link{gethr}}
}
\concept{ether functions}
