% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ether.R
\name{ether.toKwei}
\alias{ether.toKwei}
\title{Conversion to Kwei.}
\usage{
ether.toKwei(amount, type)
}
\arguments{
\item{amount}{Integer - Amount of the cryptocurrency to convert from.}

\item{type}{String - Unit of the cryptocurrency to convert from: wei, kwei,
mwei, gwei, szabo, finney, ether, kether, mether, gether or tether.}
}
\value{
Integer - Value in Kwei.
}
\description{
\code{ether.toKwei} returns the value of the cryptocurrency in Kwei.
}
\examples{
ether.toKwei(50, 'ether')

}
\seealso{
Other ether functions: \code{\link{ether.toEther}},
  \code{\link{ether.toFinney}},
  \code{\link{ether.toGether}}, \code{\link{ether.toGwei}},
  \code{\link{ether.toKether}},
  \code{\link{ether.toMether}}, \code{\link{ether.toMwei}},
  \code{\link{ether.toSzabo}},
  \code{\link{ether.toTether}}, \code{\link{ether.toWei}}
}
\concept{ether functions}
