% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ether.R
\name{ether.toTether}
\alias{ether.toTether}
\title{Conversion to Tether.}
\usage{
ether.toTether(amount, type)
}
\arguments{
\item{amount}{Integer - Amount of the cryptocurrency to convert from.}

\item{type}{String - Unit of the cryptocurrency to convert from: wei, kwei, mwei, gwei, szabo, finney, ether, kether, mether, gether or tether.}
}
\value{
Integer - Value in Tether.
}
\description{
\code{ether.toTether} returns the value of the cryptocurrency in Tether.
}
\examples{
ether.toTether(50, "ether")

}
\seealso{
Other ether functions: \code{\link{ether.toEther}},
  \code{\link{ether.toFinney}},
  \code{\link{ether.toGether}}, \code{\link{ether.toGwei}},
  \code{\link{ether.toKether}}, \code{\link{ether.toKwei}},
  \code{\link{ether.toMether}}, \code{\link{ether.toMwei}},
  \code{\link{ether.toSzabo}}, \code{\link{ether.toWei}}
}
\concept{ether functions}
