% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ether.R
\name{ether.toWei}
\alias{ether.toWei}
\title{Conversion to Wei.}
\usage{
ether.toWei(amount, type)
}
\arguments{
\item{amount}{Integer - Amount of the cryptocurrency to convert from.}

\item{type}{String - Unit of the cryptocurrency to convert from: wei, kwei, mwei, gwei, szabo, finney, ether, kether, mether, gether or tether.}
}
\value{
Integer - Value in Wei.
}
\description{
\code{ether.toWei} returns the value of the cryptocurrency in Wei.
}
\examples{
ether.toWei(50, "ether")

}
\seealso{
Other ether functions: \code{\link{ether.toEther}},
  \code{\link{ether.toFinney}},
  \code{\link{ether.toGether}}, \code{\link{ether.toGwei}},
  \code{\link{ether.toKether}}, \code{\link{ether.toKwei}},
  \code{\link{ether.toMether}}, \code{\link{ether.toMwei}},
  \code{\link{ether.toSzabo}}, \code{\link{ether.toTether}}
}
\concept{ether functions}
