% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rpc_address}
\alias{get_rpc_address}
\title{RPC address of the node.}
\usage{
get_rpc_address()
}
\value{
String - RPC Address.
}
\description{
\code{get_rpc_address} returns the RPC address that is being used to connect
to the Geth node.
}
\examples{
get_rpc_address()

}
\seealso{
Other utils functions: \code{\link{dec_to_hex}},
  \code{\link{get_post}}, \code{\link{hex_to_dec}},
  \code{\link{hex_to_text}}, \code{\link{is.wholenumber}},
  \code{\link{process_block}}, \code{\link{process_log}},
  \code{\link{process_receipt}},
  \code{\link{process_transaction}},
  \code{\link{set_rpc_address}}, \code{\link{text_to_hex}}
}
\concept{utils functions}
