% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hex_to_dec}
\alias{hex_to_dec}
\title{Hexadecimal to decimal conversion.}
\usage{
hex_to_dec(hex)
}
\arguments{
\item{hex}{Data - Value in hexadecimal.}
}
\value{
Integer - Value in decimal.
}
\description{
\code{hex_to_dec} returns the value in decimal.
}
\examples{
hex_to_dec('0x28')
hex_to_dec('0xd9190')

}
\seealso{
Other utils functions: \code{\link{dec_to_hex}},
  \code{\link{get_post}}, \code{\link{get_rpc_address}},
  \code{\link{hex_to_text}}, \code{\link{is.wholenumber}},
  \code{\link{process_block}}, \code{\link{process_log}},
  \code{\link{process_receipt}},
  \code{\link{process_transaction}},
  \code{\link{set_rpc_address}}, \code{\link{text_to_hex}}
}
\concept{utils functions}
