% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hex_to_text}
\alias{hex_to_text}
\title{Hexadecimal to string conversion.}
\usage{
hex_to_text(msg)
}
\arguments{
\item{msg}{Data - Value in hexadecimal.}
}
\value{
String - Value in plain text.
}
\description{
\code{hex_to_text} returns the plain text.
}
\examples{
hex_to_text("0x68656c6c6f20776f726c64")

}
\seealso{
Other utils functions: \code{\link{dec_to_hex}},
  \code{\link{get_post}}, \code{\link{get_rpc_address}},
  \code{\link{hex_to_dec}}, \code{\link{is.wholenumber}},
  \code{\link{process_block}}, \code{\link{process_log}},
  \code{\link{process_receipt}},
  \code{\link{process_transaction}},
  \code{\link{set_rpc_address}}, \code{\link{text_to_hex}}
}
\concept{utils functions}
