% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Whole numbers identification.}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{Data - Element to check whether it is a whole number.}

\item{tol}{Double - Range to determine if the number is a whole number.}
}
\value{
Boolean - \code{true} if the input is a whole number, \code{false} otherwise.
}
\description{
\code{is.wholenumber} returns whether a value is a whole number.
}
\examples{
is.wholenumber(20)
is.wholenumber(20.3)
is.wholenumber("no")

}
\seealso{
Other utils functions: \code{\link{dec_to_hex}},
  \code{\link{get_post}}, \code{\link{get_rpc_address}},
  \code{\link{hex_to_dec}}, \code{\link{hex_to_text}},
  \code{\link{process_block}}, \code{\link{process_log}},
  \code{\link{process_receipt}},
  \code{\link{process_transaction}},
  \code{\link{set_rpc_address}}, \code{\link{text_to_hex}}
}
\concept{utils functions}
