% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miner.R
\name{miner_setEtherBase}
\alias{miner_setEtherBase}
\title{Etherbase for mining.}
\usage{
miner_setEtherBase(address)
}
\arguments{
\item{address}{Address - Address to send the rewards when mining.}
}
\value{
Boolean - \code{true} if changed, otherwise \code{false}.
}
\description{
\code{miner_setEtherBase} sets the etherbase, where mining rewards will go.
}
\examples{
\donttest{
miner_setEtherBase("0xb117a8bc3ecf2c3f006b89da6826e49b4193977a")
}

}
\seealso{
Other miner functions: \code{\link{miner_setExtra}},
  \code{\link{miner_setGasPrice}},
  \code{\link{miner_start}}, \code{\link{miner_stop}}
}
\concept{miner functions}
