% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miner.R
\name{miner_setExtra}
\alias{miner_setExtra}
\title{Extra data for mined blocks.}
\usage{
miner_setExtra(data)
}
\arguments{
\item{data}{String - Data to include in mined blocks.}
}
\value{
Boolean - \code{true} if changed, otherwise \code{false}.
}
\description{
\code{miner_setExtra} sets the extra data a miner can include when miner blocks. This is capped at 32 bytes.
}
\examples{
\donttest{
miner_setExtra("My info")
}

}
\seealso{
Other miner functions: \code{\link{miner_setEtherBase}},
  \code{\link{miner_setGasPrice}},
  \code{\link{miner_start}}, \code{\link{miner_stop}}
}
\concept{miner functions}
