% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miner.R
\name{miner_start}
\alias{miner_start}
\title{Mining process to be started.}
\usage{
miner_start(threads = 1)
}
\arguments{
\item{threads}{Integer - Given number of threads used for mining.}
}
\description{
\code{miner_start} starts the CPU mining process with the given number of
threads.
}
\examples{
\donttest{
miner_start()
miner_start(4)
}

}
\seealso{
Other miner functions: \code{\link{gethr}},
  \code{\link{miner_setEtherBase}},
  \code{\link{miner_setExtra}},
  \code{\link{miner_setGasPrice}}, \code{\link{miner_stop}}
}
\concept{miner functions}
