% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miner.R
\name{miner_stop}
\alias{miner_stop}
\title{Mining process to be stopped.}
\usage{
miner_stop()
}
\description{
\code{miner_stop} stops the CPU mining operation.
}
\examples{
\donttest{
miner_stop()
}

}
\seealso{
Other miner functions: \code{\link{gethr}},
  \code{\link{miner_setEtherBase}},
  \code{\link{miner_setExtra}},
  \code{\link{miner_setGasPrice}},
  \code{\link{miner_start}}
}
\concept{miner functions}
