% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{net_listening}
\alias{net_listening}
\title{Whether client is listening or not.}
\usage{
net_listening()
}
\value{
Boolean - \code{true} when listening, otherwise \code{false}.
}
\description{
\code{net_listening} returns \code{true} if client is actively listening for
network connections.
}
\examples{
\donttest{
net_listening()
}

}
\seealso{
Other net functions: \code{\link{gethr}},
  \code{\link{net_peerCount}}, \code{\link{net_version}}
}
\concept{net functions}
