% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{net_peerCount}
\alias{net_peerCount}
\title{Number of peers connected.}
\usage{
net_peerCount()
}
\value{
Integer - Number of connected peers.
}
\description{
\code{net_peerCount} returns the number of peers currently connected to the
client.
}
\examples{
\donttest{
net_peerCount()
}

}
\seealso{
Other net functions: \code{\link{gethr}},
  \code{\link{net_listening}}, \code{\link{net_version}}
}
\concept{net functions}
