% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{net_version}
\alias{net_version}
\title{Current network id.}
\usage{
net_version()
}
\value{
String - Current network id.
}
\description{
\code{net_version} returns the current network id.
}
\examples{
\donttest{
net_version()
}

}
\seealso{
Other net functions: \code{\link{net_listening}},
  \code{\link{net_peerCount}}
}
\concept{net functions}
