% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personal.R
\name{personal_ecRecover}
\alias{personal_ecRecover}
\title{Signatory address.}
\usage{
personal_ecRecover(message, signature)
}
\arguments{
\item{message}{String - Message that has been signed.}

\item{signature}{Data - Signature of the message that has been signed.}
}
\value{
Address - Address of the account that has signed a message.
}
\description{
\code{personal_ecRecover} returns the address associated with the private key
that was used to calculate a signature.
}
\examples{
\donttest{
personal_ecRecover('hello world', '0x1dd3657c91d95f350ab25f17ee7cbcdbccd3f5bc
52976bfd4dd03bd6bc29d2ac23e656bee509ca33b921e0e6b53eb64082be1bb3c69c3a4adccd9
93b1d667f8d1b')
}

}
\seealso{
Other personal functions: \code{\link{personal_importRawKey}},
  \code{\link{personal_listAccounts}},
  \code{\link{personal_lockAccount}},
  \code{\link{personal_newAccount}},
  \code{\link{personal_sign}},
  \code{\link{personal_unlockAccount}}
}
\concept{personal functions}
