% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personal.R
\name{personal_importRawKey}
\alias{personal_importRawKey}
\title{New account creation giving the private key.}
\usage{
personal_importRawKey(keydata, password)
}
\arguments{
\item{keydata}{Data - Message that has been signed.}

\item{password}{String - Password of the account.}
}
\value{
Address - Address of the new account.
}
\description{
\code{personal_importRawKey} imports the given unencrypted private key (hex
string) into the key store, encrypting it with the passphrase.
}
\examples{
\donttest{
personal_importRawKey('a5e3d0b2bb3011d00a139e5cdc4ae13144962752d6af7916bf2b
d271a240094e', 'password')
}

}
\seealso{
Other personal functions: \code{\link{gethr}},
  \code{\link{personal_ecRecover}},
  \code{\link{personal_listAccounts}},
  \code{\link{personal_lockAccount}},
  \code{\link{personal_newAccount}},
  \code{\link{personal_sign}},
  \code{\link{personal_unlockAccount}}
}
\concept{personal functions}
