% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personal.R
\name{personal_listAccounts}
\alias{personal_listAccounts}
\title{Addresses owned by client.}
\usage{
personal_listAccounts()
}
\value{
Array of Address - Addresses owned by the client.
}
\description{
\code{personal_lockAccount} returns a list of addresses owned by client.
}
\examples{
\donttest{
personal_listAccounts()
}

}
\seealso{
Other personal functions: \code{\link{gethr}},
  \code{\link{personal_ecRecover}},
  \code{\link{personal_importRawKey}},
  \code{\link{personal_lockAccount}},
  \code{\link{personal_newAccount}},
  \code{\link{personal_sign}},
  \code{\link{personal_unlockAccount}}
}
\concept{personal functions}
