% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personal.R
\name{personal_lockAccount}
\alias{personal_lockAccount}
\title{Account deletion.}
\usage{
personal_lockAccount(address)
}
\arguments{
\item{address}{Address - Address of account to be deleted.}
}
\value{
Boolean - \code{true} if the account has been deleted.
}
\description{
\code{personal_lockAccount} removes the private key with given address from memory. The account can no longer be used to send transactions.
}
\examples{
\donttest{
personal_lockAccount("0xf1b76d9a65b532dbdc3899dee6e117b52c85a536")
}

}
\seealso{
Other personal functions: \code{\link{personal_ecRecover}},
  \code{\link{personal_importRawKey}},
  \code{\link{personal_listAccounts}},
  \code{\link{personal_newAccount}},
  \code{\link{personal_sign}},
  \code{\link{personal_unlockAccount}}
}
\concept{personal functions}
