% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personal.R
\name{personal_newAccount}
\alias{personal_newAccount}
\title{New account creation.}
\usage{
personal_newAccount(password)
}
\arguments{
\item{password}{String - Password of the account.}
}
\value{
Address - Address of the new account.
}
\description{
\code{personal_newAccount} generates a new private key and stores it in the
key store directory. The key file is encrypted with the given passphrase.
}
\examples{
\donttest{
personal_newAccount('password')
}

}
\seealso{
Other personal functions: \code{\link{gethr}},
  \code{\link{personal_ecRecover}},
  \code{\link{personal_importRawKey}},
  \code{\link{personal_listAccounts}},
  \code{\link{personal_lockAccount}},
  \code{\link{personal_sign}},
  \code{\link{personal_unlockAccount}}
}
\concept{personal functions}
