% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personal.R
\name{personal_sign}
\alias{personal_sign}
\title{Ethereum specific signature}
\usage{
personal_sign(address, message, password)
}
\arguments{
\item{address}{Address - Address of the user.}

\item{message}{String - Message to sign.}

\item{password}{String - Password of the account.}
}
\value{
Data - Signature.
}
\description{
\code{personal_sign} returns an Ethereum specific signature with
\code{keccak256}.
}
\examples{
\donttest{
personal_sign('0xb117a8bc3ecf2c3f006b89da6826e49b4193977a', 'hello world',
'password')
}

}
\seealso{
Other personal functions: \code{\link{gethr}},
  \code{\link{personal_ecRecover}},
  \code{\link{personal_importRawKey}},
  \code{\link{personal_listAccounts}},
  \code{\link{personal_lockAccount}},
  \code{\link{personal_newAccount}},
  \code{\link{personal_unlockAccount}}
}
\concept{personal functions}
