% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personal.R
\name{personal_unlockAccount}
\alias{personal_unlockAccount}
\title{Unlocking the key.}
\usage{
personal_unlockAccount(address, password, duration = 3000)
}
\arguments{
\item{address}{Address - Address to unlock.}

\item{password}{String - Password of the account.}

\item{duration}{Integer - Seconds the account will be unlocked.}
}
\value{
Boolean - \code{true} if the account has been unlocked.
}
\description{
\code{personal_unlockAccount} unlocks an account to be used for signing and
transactions.
}
\examples{
\donttest{
personal_unlockAccount('0xb117a8bc3ecf2c3f006b89da6826e49b4193977a',
'password', 60000)
}

}
\seealso{
Other personal functions: \code{\link{personal_ecRecover}},
  \code{\link{personal_importRawKey}},
  \code{\link{personal_listAccounts}},
  \code{\link{personal_lockAccount}},
  \code{\link{personal_newAccount}},
  \code{\link{personal_sign}}
}
\concept{personal functions}
