% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_transaction}
\alias{process_transaction}
\title{Values of the transactions in plain text or decimal instead of hexadecimal.}
\usage{
process_transaction(trans)
}
\arguments{
\item{trans}{Objects - Information of the transaction in hexadecimal.}
}
\value{
Object - Information of the transaction with values in plain text or
  decimal if possible.
}
\description{
\code{process_block} returns the values of the transactions in plain text or
decimal if possible. The values are obtained from the Geth node.
}
\seealso{
Other utils functions: \code{\link{dec_to_hex}},
  \code{\link{get_post}}, \code{\link{get_rpc_address}},
  \code{\link{hex_to_dec}}, \code{\link{hex_to_text}},
  \code{\link{is.wholenumber}},
  \code{\link{process_block}}, \code{\link{process_log}},
  \code{\link{process_receipt}},
  \code{\link{set_rpc_address}}, \code{\link{text_to_hex}}
}
\concept{utils functions}
