% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_rpc_address}
\alias{set_rpc_address}
\title{Query of the RPC address of the node.}
\usage{
set_rpc_address(url, port)
}
\arguments{
\item{url}{String - URL of the Geth node.}

\item{port}{Integer - Port of the Geth node.}
}
\description{
\code{set_rpc_address} sets the RPC address that is being used to connect to
the Geth node.
}
\examples{
set_rpc_address('http://153.35.91.1', 8600)
set_rpc_address('http://localhost', 8545)

}
\seealso{
Other utils functions: \code{\link{dec_to_hex}},
  \code{\link{get_post}}, \code{\link{get_rpc_address}},
  \code{\link{hex_to_dec}}, \code{\link{hex_to_text}},
  \code{\link{is.wholenumber}},
  \code{\link{process_block}}, \code{\link{process_log}},
  \code{\link{process_receipt}},
  \code{\link{process_transaction}},
  \code{\link{text_to_hex}}
}
\concept{utils functions}
