% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shh.R
\name{shh_addPrivatekey}
\alias{shh_addPrivatekey}
\title{Key pair storage.}
\usage{
shh_addPrivatekey(key)
}
\arguments{
\item{key}{String - Key as HEX bytes.}
}
\value{
String - Key pair ID on sucess.
}
\description{
\code{shh_addPrivatekey} stores the key pair, and returns its ID.
}
\examples{
\donttest{
shh_addPrivatekey("0x6d795f707269766174655f6b6579")
}

}
\seealso{
Other shh functions: \code{\link{shh_addSymKey}},
  \code{\link{shh_deleteKeyPair}},
  \code{\link{shh_deleteSymKey}},
  \code{\link{shh_generateSymKeyFromPassword}},
  \code{\link{shh_getPrivateKey}},
  \code{\link{shh_getPublicKey}},
  \code{\link{shh_getSymKey}},
  \code{\link{shh_hasKeyPair}},
  \code{\link{shh_hasSymKey}}, \code{\link{shh_info}},
  \code{\link{shh_markTrustedPeer}},
  \code{\link{shh_newKeyPair}},
  \code{\link{shh_newMessageFilter}},
  \code{\link{shh_newSymKey}}, \code{\link{shh_post}},
  \code{\link{shh_setMaxMessageSize}},
  \code{\link{shh_setMinPoW}}, \code{\link{shh_version}}
}
\concept{shh functions}
