% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shh.R
\name{shh_addSymKey}
\alias{shh_addSymKey}
\title{Symmetric key storage.}
\usage{
shh_addSymKey(key)
}
\arguments{
\item{key}{String - Key as HEX bytes.}
}
\value{
String - Symmetric key ID on sucess.
}
\description{
\code{shh_addSymKey} stores the symmetric key, and returns its ID.
}
\examples{
\donttest{
shh_addSymKey('0xf6dcf21ed6a17bd78d8c4c63195ab997b3b65ea683705501eae82d32667adc92')
}

}
\seealso{
Other shh functions: \code{\link{gethr}},
  \code{\link{shh_addPrivatekey}},
  \code{\link{shh_deleteKeyPair}},
  \code{\link{shh_deleteSymKey}},
  \code{\link{shh_generateSymKeyFromPassword}},
  \code{\link{shh_getPrivateKey}},
  \code{\link{shh_getPublicKey}},
  \code{\link{shh_getSymKey}},
  \code{\link{shh_hasKeyPair}},
  \code{\link{shh_hasSymKey}}, \code{\link{shh_info}},
  \code{\link{shh_markTrustedPeer}},
  \code{\link{shh_newKeyPair}},
  \code{\link{shh_newMessageFilter}},
  \code{\link{shh_newSymKey}}, \code{\link{shh_post}},
  \code{\link{shh_setMaxMessageSize}},
  \code{\link{shh_setMinPoW}}, \code{\link{shh_version}}
}
\concept{shh functions}
