% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shh.R
\name{shh_deleteKeyPair}
\alias{shh_deleteKeyPair}
\title{Key pair deletion.}
\usage{
shh_deleteKeyPair(id)
}
\arguments{
\item{id}{String - ID of the key pair.}
}
\value{
Boolean - \code{true} on success.
}
\description{
\code{shh_addPrivatekey} deletes the specifies key pair if it exists.
}
\examples{
\donttest{
shh_deleteKeyPair('7ade0794bf59b9a4508a3c5d7d3408f910fc575fb2f31845da7611cae6664448')
}

}
\seealso{
Other shh functions: \code{\link{gethr}},
  \code{\link{shh_addPrivatekey}},
  \code{\link{shh_addSymKey}},
  \code{\link{shh_deleteSymKey}},
  \code{\link{shh_generateSymKeyFromPassword}},
  \code{\link{shh_getPrivateKey}},
  \code{\link{shh_getPublicKey}},
  \code{\link{shh_getSymKey}},
  \code{\link{shh_hasKeyPair}},
  \code{\link{shh_hasSymKey}}, \code{\link{shh_info}},
  \code{\link{shh_markTrustedPeer}},
  \code{\link{shh_newKeyPair}},
  \code{\link{shh_newMessageFilter}},
  \code{\link{shh_newSymKey}}, \code{\link{shh_post}},
  \code{\link{shh_setMaxMessageSize}},
  \code{\link{shh_setMinPoW}}, \code{\link{shh_version}}
}
\concept{shh functions}
