% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shh.R
\name{shh_deleteSymKey}
\alias{shh_deleteSymKey}
\title{Symmetric key deletion.}
\usage{
shh_deleteSymKey(id)
}
\arguments{
\item{id}{String - ID of the symmetric key.}
}
\value{
Boolean - \code{true} on success.
}
\description{
\code{shh_deleteSymKey} deletes the specifies symmetric key if it exists.
}
\examples{
\donttest{
shh_deleteSymKey("8d7b2dff569d14308a8e74ca1475dd93ba8dd42a9a74e97638796d5d6c8751ac")
}

}
\seealso{
Other shh functions: \code{\link{shh_addPrivatekey}},
  \code{\link{shh_addSymKey}},
  \code{\link{shh_deleteKeyPair}},
  \code{\link{shh_generateSymKeyFromPassword}},
  \code{\link{shh_getPrivateKey}},
  \code{\link{shh_getPublicKey}},
  \code{\link{shh_getSymKey}},
  \code{\link{shh_hasKeyPair}},
  \code{\link{shh_hasSymKey}}, \code{\link{shh_info}},
  \code{\link{shh_markTrustedPeer}},
  \code{\link{shh_newKeyPair}},
  \code{\link{shh_newMessageFilter}},
  \code{\link{shh_newSymKey}}, \code{\link{shh_post}},
  \code{\link{shh_setMaxMessageSize}},
  \code{\link{shh_setMinPoW}}, \code{\link{shh_version}}
}
\concept{shh functions}
