% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shh.R
\name{shh_hasSymKey}
\alias{shh_hasSymKey}
\title{Symmetric key existence confirmation.}
\usage{
shh_hasSymKey(id)
}
\arguments{
\item{id}{String - ID of the symmetric key.}
}
\value{
Boolean - \code{true} if the symmetric key exists, otherwise
  \code{false}.
}
\description{
\code{shh_hasSymKey} checks if the whisper node has a symmetric key given an
ID.
}
\examples{
\donttest{
shh_hasSymKey('4c5e37602b120904c592181e75917949c07b0d2f3111d41bb554b604e532c197')
}

}
\seealso{
Other shh functions: \code{\link{shh_addPrivatekey}},
  \code{\link{shh_addSymKey}},
  \code{\link{shh_deleteKeyPair}},
  \code{\link{shh_deleteSymKey}},
  \code{\link{shh_generateSymKeyFromPassword}},
  \code{\link{shh_getPrivateKey}},
  \code{\link{shh_getPublicKey}},
  \code{\link{shh_getSymKey}},
  \code{\link{shh_hasKeyPair}}, \code{\link{shh_info}},
  \code{\link{shh_markTrustedPeer}},
  \code{\link{shh_newKeyPair}},
  \code{\link{shh_newMessageFilter}},
  \code{\link{shh_newSymKey}}, \code{\link{shh_post}},
  \code{\link{shh_setMaxMessageSize}},
  \code{\link{shh_setMinPoW}}, \code{\link{shh_version}}
}
\concept{shh functions}
