% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shh.R
\name{shh_info}
\alias{shh_info}
\title{Information about the whisper.}
\usage{
shh_info()
}
\value{
Object - Diagnostic information about current minimum PoW
  requirement, message size limit in bytes, memory size of the messages and
  number of current messages.
}
\description{
\code{shh_info} returns diagnostic information about the whisper node.
}
\examples{
\donttest{
shh_info()
}

}
\seealso{
Other shh functions: \code{\link{shh_addPrivatekey}},
  \code{\link{shh_addSymKey}},
  \code{\link{shh_deleteKeyPair}},
  \code{\link{shh_deleteSymKey}},
  \code{\link{shh_generateSymKeyFromPassword}},
  \code{\link{shh_getPrivateKey}},
  \code{\link{shh_getPublicKey}},
  \code{\link{shh_getSymKey}},
  \code{\link{shh_hasKeyPair}},
  \code{\link{shh_hasSymKey}},
  \code{\link{shh_markTrustedPeer}},
  \code{\link{shh_newKeyPair}},
  \code{\link{shh_newMessageFilter}},
  \code{\link{shh_newSymKey}}, \code{\link{shh_post}},
  \code{\link{shh_setMaxMessageSize}},
  \code{\link{shh_setMinPoW}}, \code{\link{shh_version}}
}
\concept{shh functions}
