% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shh.R
\name{shh_newKeyPair}
\alias{shh_newKeyPair}
\title{Key pair creation and storage.}
\usage{
shh_newKeyPair()
}
\value{
String - Key pair ID on sucess.
}
\description{
\code{shh_newKeyPair} generates a new public and private key pair for message decryption and encryption.
}
\examples{
\donttest{
shh_newKeyPair()
}

}
\seealso{
Other shh functions: \code{\link{shh_addPrivatekey}},
  \code{\link{shh_addSymKey}},
  \code{\link{shh_deleteKeyPair}},
  \code{\link{shh_deleteSymKey}},
  \code{\link{shh_generateSymKeyFromPassword}},
  \code{\link{shh_getPrivateKey}},
  \code{\link{shh_getPublicKey}},
  \code{\link{shh_getSymKey}},
  \code{\link{shh_hasKeyPair}},
  \code{\link{shh_hasSymKey}}, \code{\link{shh_info}},
  \code{\link{shh_markTrustedPeer}},
  \code{\link{shh_newMessageFilter}},
  \code{\link{shh_newSymKey}}, \code{\link{shh_post}},
  \code{\link{shh_setMaxMessageSize}},
  \code{\link{shh_setMinPoW}}, \code{\link{shh_version}}
}
\concept{shh functions}
