% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shh.R
\name{shh_newSymKey}
\alias{shh_newSymKey}
\title{Symmetric key creation and storage.}
\usage{
shh_newSymKey()
}
\value{
String - Symmetric key ID on sucess.
}
\description{
\code{shh_newSymKey} generates a random symmetric key and stores it under an
ID, which is then returned. Can be used encrypting and decrypting messages
where the key is known to both parties.
}
\examples{
\donttest{
shh_newSymKey()
}

}
\seealso{
Other shh functions: \code{\link{gethr}},
  \code{\link{shh_addPrivatekey}},
  \code{\link{shh_addSymKey}},
  \code{\link{shh_deleteKeyPair}},
  \code{\link{shh_deleteSymKey}},
  \code{\link{shh_generateSymKeyFromPassword}},
  \code{\link{shh_getPrivateKey}},
  \code{\link{shh_getPublicKey}},
  \code{\link{shh_getSymKey}},
  \code{\link{shh_hasKeyPair}},
  \code{\link{shh_hasSymKey}}, \code{\link{shh_info}},
  \code{\link{shh_markTrustedPeer}},
  \code{\link{shh_newKeyPair}},
  \code{\link{shh_newMessageFilter}},
  \code{\link{shh_post}},
  \code{\link{shh_setMaxMessageSize}},
  \code{\link{shh_setMinPoW}}, \code{\link{shh_version}}
}
\concept{shh functions}
