% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shh.R
\name{shh_setMinPoW}
\alias{shh_setMinPoW}
\title{Minimal Pow.}
\usage{
shh_setMinPoW(pow)
}
\arguments{
\item{pow}{Integer - The new PoW requirement.}
}
\value{
Boolean -  \code{true} on success.
}
\description{
\code{shh_setMinPoW} sets the minimal PoW required by this node.
}
\examples{
\donttest{
shh_setMinPoW(0.2)
}

}
\seealso{
Other shh functions: \code{\link{gethr}},
  \code{\link{shh_addPrivatekey}},
  \code{\link{shh_addSymKey}},
  \code{\link{shh_deleteKeyPair}},
  \code{\link{shh_deleteSymKey}},
  \code{\link{shh_generateSymKeyFromPassword}},
  \code{\link{shh_getPrivateKey}},
  \code{\link{shh_getPublicKey}},
  \code{\link{shh_getSymKey}},
  \code{\link{shh_hasKeyPair}},
  \code{\link{shh_hasSymKey}}, \code{\link{shh_info}},
  \code{\link{shh_markTrustedPeer}},
  \code{\link{shh_newKeyPair}},
  \code{\link{shh_newMessageFilter}},
  \code{\link{shh_newSymKey}}, \code{\link{shh_post}},
  \code{\link{shh_setMaxMessageSize}},
  \code{\link{shh_version}}
}
\concept{shh functions}
