% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shh.R
\name{shh_version}
\alias{shh_version}
\title{Current whisper protocol.}
\usage{
shh_version()
}
\value{
String - Current whisper protocol version.
}
\description{
\code{shh_version} returns the current whisper protocol version.
}
\examples{
\donttest{
shh_version()
}

}
\seealso{
Other shh functions: \code{\link{shh_addPrivatekey}},
  \code{\link{shh_addSymKey}},
  \code{\link{shh_deleteKeyPair}},
  \code{\link{shh_deleteSymKey}},
  \code{\link{shh_generateSymKeyFromPassword}},
  \code{\link{shh_getPrivateKey}},
  \code{\link{shh_getPublicKey}},
  \code{\link{shh_getSymKey}},
  \code{\link{shh_hasKeyPair}},
  \code{\link{shh_hasSymKey}}, \code{\link{shh_info}},
  \code{\link{shh_markTrustedPeer}},
  \code{\link{shh_newKeyPair}},
  \code{\link{shh_newMessageFilter}},
  \code{\link{shh_newSymKey}}, \code{\link{shh_post}},
  \code{\link{shh_setMaxMessageSize}},
  \code{\link{shh_setMinPoW}}
}
\concept{shh functions}
