% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{text_to_hex}
\alias{text_to_hex}
\title{String to hexadecimal conversion.}
\usage{
text_to_hex(text)
}
\arguments{
\item{text}{String - Value in plain text.}
}
\value{
Data - Value in hexadecimal.
}
\description{
\code{text_to_hex} returns the hexadecimal.
}
\examples{
text_to_hex("hello world")

}
\seealso{
Other utils functions: \code{\link{dec_to_hex}},
  \code{\link{get_post}}, \code{\link{get_rpc_address}},
  \code{\link{hex_to_dec}}, \code{\link{hex_to_text}},
  \code{\link{is.wholenumber}},
  \code{\link{process_block}}, \code{\link{process_log}},
  \code{\link{process_receipt}},
  \code{\link{process_transaction}},
  \code{\link{set_rpc_address}}
}
\concept{utils functions}
