% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txpool.R
\name{txpool_content}
\alias{txpool_content}
\title{Information about pending and queued transactions.}
\usage{
txpool_content()
}
\value{
Object - Information about the pending and queued transactions to be
  inserted in the blockchain.
}
\description{
\code{txpool_content} returns the list the exact details of all the
transactions currently pending for inclusion in the next block(s), as well as
the ones that are being scheduled for future execution only.
}
\examples{
\donttest{
txpool_content()
}

}
\seealso{
Other txpool functions: \code{\link{txpool_inspect}},
  \code{\link{txpool_status}}
}
\concept{txpool functions}
