% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txpool.R
\name{txpool_inspect}
\alias{txpool_inspect}
\title{Summary of the information about pending and queued transactions.}
\usage{
txpool_inspect()
}
\value{
Object - Summary of the information about the pending and queued
  transactions to be inserted in the blockchain.
}
\description{
\code{txpool_inspect} returns the list a textual summary of all the
transactions currently pending for inclusion in the next block(s), as well as
the ones that are being scheduled for future execution only. This is a method
specifically tailored to developers to quickly see the transactions in the
pool and find any potential issues.
}
\examples{
\donttest{
txpool_inspect()
}

}
\seealso{
Other txpool functions: \code{\link{gethr}},
  \code{\link{txpool_content}}, \code{\link{txpool_status}}
}
\concept{txpool functions}
