% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txpool.R
\name{txpool_status}
\alias{txpool_status}
\title{Number of pending and queued transactions.}
\usage{
txpool_status()
}
\value{
Integer - Number of pending and queued transactions.
}
\description{
\code{txpool_status} returns the number of transactions currently pending for
inclusion in the next block(s), as well as the ones that are being scheduled
for future execution only.
}
\examples{
\donttest{
txpool_status()
}

}
\seealso{
Other txpool functions: \code{\link{gethr}},
  \code{\link{txpool_content}},
  \code{\link{txpool_inspect}}
}
\concept{txpool functions}
