% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web3.R
\name{web3_clientVersion}
\alias{web3_clientVersion}
\title{Current client version.}
\usage{
web3_clientVersion()
}
\value{
String - Current client version.
}
\description{
\code{web3_clientVersion} returns the current client version.
}
\examples{
\donttest{
web3_clientVersion()
}

}
\seealso{
Other web3 functions: \code{\link{web3_sha3}}
}
\concept{web3 functions}
