% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web3.R
\name{web3_sha3}
\alias{web3_sha3}
\title{Keccak-256 value of the data.}
\usage{
web3_sha3(data)
}
\arguments{
\item{data}{String - Data to convert into a SHA3 hash.}
}
\value{
Data - SHA3 result of the given string.
}
\description{
\code{web3_sha3} returns Keccak-256 (not the standardized SHA3-256) of the
given data.
}
\examples{
\donttest{
web3_sha3('hello')
}

}
\seealso{
Other web3 functions: \code{\link{web3_clientVersion}}
}
\concept{web3 functions}
